In-progress commit with NodeData.
This feels like closer to a final solution, where the Node is a trait
with a required get_node_data() method and default implementations fo
get_x(), get_name() etc. that refer to the NodeData object.

However we're still fighting the borrow checker as we need references to
the input and output ports in the NodeData, while keeping the owned
Ports in the concrete node implementation. Not sure how to do that
properly.