Simplify node / function setup.
Nodes are now the main data structure. They carry arround the function.
When rendering, the function gets invoked with the node and rendering
context. Functions don't store state locally (even though they could),
but only through the render context. The function is responsible for
setting up the node: adding ports, etc.

The main "trick" here to get around the borrow checker is to initialize
the node with a dummy function first, then create the actual function
and do the setup, then finally do a std::mem::swap to swap in the
correct function with the dummy.